Change max_tries/max_simultaneous from u32 to i32 in API types and config
Unify integer types so the API-facing types (AgentRecord, UpsertAgentRequest)
and config constants match the domain Agent struct (i32), eliminating all
as-i32/as-u32 boundary casts in routes/agents.rs, spawner.rs, and
run_spawners.rs.

+++ metis-server/src/background/run_spawners.rs
@@ -166,8 +166,8 @@ mod tests {
         let agent = Agent::new(
             name.to_string(),
             format!("/agents/{name}/prompt.md"),
-            DEFAULT_AGENT_MAX_TRIES as i32,
-            DEFAULT_AGENT_MAX_SIMULTANEOUS as i32,
+            DEFAULT_AGENT_MAX_TRIES,
+            DEFAULT_AGENT_MAX_SIMULTANEOUS,
             false,
         );
         handles.store.add_agent(agent).await?;
